# Clean up `mass_matrix!` and `constraint_jacobian!`