* Add `*(::Transpose(A), ::B)` overloads involving `WrenchMatrix`, `MomentumMatrix`, and `GeometricJacobian`
* Use these new overloads in `mass_matrix!` and `constraint_jacobian!`
* add `supports` function and use it to further clean up `mass_matrix!`
* Iterate directly over joint velocity indices in `mass_matrix!` and look up their corresponding `JointID`s, instead of first iterating over `JointIDs` and then over their `velocity_range`s.

The last point above results in a bit of a speedup for `mass_matrix!`:

Before:

```
  minimum time:     5.930 μs (0.00% GC)
  median time:      6.028 μs (0.00% GC)
  mean time:        6.211 μs (0.00% GC)
  maximum time:     9.859 μs (0.00% GC)
```

After:

```
  minimum time:     5.435 μs (0.00% GC)
  median time:      5.546 μs (0.00% GC)
  mean time:        5.741 μs (0.00% GC)
  maximum time:     9.264 μs (0.00% GC)
```